added if condition to mail list

+++ mail2.php
@@ -6,15 +6,32 @@
   $to=$_POST['email'];
   $subject="Unsubscribed!";
   $body ="You've unsubscribed from Sa3y successfully! we hope to see again :)";
-  mail($to,$subject,$body);
-  $sql = "DELETE FROM mail WHERE email='$to'";
-  mysqli_query($db, $sql);
-?>
+  
+  $sql1 = "select * from mail where email='$to' ";
+  $result = mysqli_query($db, $sql1);
+  $count = mysqli_num_rows($result);
+ 
+	
+  if ($count == 0) {
+		?>
+		
+        <script type="text/javascript">
+		alert("لم تسجل في القائمة البريدية بعد!");
+		location="index.html";
+		</script>
+        
+        <?php
+  } else { 
+      mail($to,$subject,$body);
+      $sql = "DELETE FROM mail WHERE email='$to'";
+     mysqli_query($db, $sql);
+         ?>
+  
   <script type="text/javascript">
-   alert(" تمت إزالتك من القائمة البريدية بنجاح!");
+   alert("تم إلغاء اشتراكك بالقائمة البريدية بنجاح!");
    location="unsubConfirm.html";
    </script>
    <?php	
   include("unsubConfirm.html");
-  }
+  }}
 ?>
